requirements-gcs.txt
google-cloud-storage>=2.16.0,<3

